Refactor quiz analysis controller tests
